% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cik.to.10}
\alias{cik.to.10}
\title{CIK to 10 digit}
\usage{
cik.to.10(cik)
}
\arguments{
\item{cik}{CIK number}
}
\value{
The 10-digit cik version.
}
\description{
\code{cik.to.10} Convert cik to 10 digit.
}
\details{
\@usage cik.to.10(cik)
}
\examples{
\dontrun{

cik.to.10(1321655)
}
}
