% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompanyConcept.R
\name{getCompanyConcept}
\alias{getCompanyConcept}
\title{Using the EDGAR API, download a list of submissions by this company.}
\usage{
getCompanyConcept(cik, concept)
}
\arguments{
\item{cik}{the cik number of the firm.}

\item{concept}{a taxonomy and tag in the form of 'taxonomy/tag'}
}
\value{
dataframe list of all submitted reports by the firm.
}
\description{
\code{getCompanyConcept} retrieve the list of reports submitted by this company.
}
\details{
getCompanyConcept takes the cik of the firm as an input parameter and provides
information about the reports submitted to EDGAR by the firm.
}
\examples{
\dontrun{
  pltr.df <- getCompanyConcept(1321655, 'us-gaap/tag')
}
}
