% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubmissions.R
\name{getSubmissions}
\alias{getSubmissions}
\title{Using the EDGAR API, download a list of submissions by this company.}
\usage{
getSubmissions(cik)
}
\arguments{
\item{cik}{the cik number of the firm.}
}
\value{
dataframe list of all submitted reports by the firm.
}
\description{
\code{getSubmissions} retrieve the list of reports submitted by this company.
}
\details{
getSubmissions takes the cik of the firm as an input parameter and provides
information about the reports submitted to EDGAR by the firm.
}
\examples{
\dontrun{
  pltr.df <- getSubmissions(1321655)
}
}
