% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setUserAgent}
\alias{setUserAgent}
\title{Set User Agent}
\usage{
setUserAgent(ua)
}
\arguments{
\item{ua}{user agent string.}
}
\value{
NONE
}
\description{
\code{setUserAgent} Set user agent header.
}
\examples{
\dontrun{

setUserAgent('jbennett@jbennettconsulting.com')
}
}
